Rename `services` -> `service`
