+++ api/src/main/java/me/lokka30/treasury/api/common/service/Service.java
@@ -2,7 +2,7 @@
  * This file is/was part of Treasury. To read more information about Treasury such as its licensing, see <https://github.com/lokka30/Treasury>.
  */
 
-package me.lokka30.treasury.api.common.services;
+package me.lokka30.treasury.api.common.service;
 
 import java.util.Objects;
 import java.util.function.Supplier;
