+++ api/src/main/java/me/lokka30/treasury/api/common/service/event/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Holds the services events
+ */
+package me.lokka30.treasury.api.common.service.event;
