+++ api/src/main/java/me/lokka30/treasury/api/common/service/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Holds classes about services and service providing.
  */
-package me.lokka30.treasury.api.common.services;
+package me.lokka30.treasury.api.common.service;
