--- api/src/main/java/me/lokka30/treasury/api/common/services/event/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Holds the services events
- */
-package me.lokka30.treasury.api.common.services.event;
