+++ platform/bukkit/plugin/src/main/java/me/lokka30/treasury/plugin/bukkit/TreasuryBukkit.java
@@ -10,8 +10,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.logging.Filter;
 import me.lokka30.treasury.api.common.event.EventExecutorTrackerShutdown;
-import me.lokka30.treasury.api.common.services.Service;
-import me.lokka30.treasury.api.common.services.ServiceProvider;
+import me.lokka30.treasury.api.common.service.Service;
+import me.lokka30.treasury.api.common.service.ServiceProvider;
 import me.lokka30.treasury.api.economy.EconomyProvider;
 import me.lokka30.treasury.api.economy.misc.OptionalEconomyApiFeature;
 import me.lokka30.treasury.plugin.bukkit.command.TreasuryCommand;
