+++ platform/bukkit/plugin/src/main/java/me/lokka30/treasury/plugin/bukkit/services/treasury2bukkit/T2BServiceMigrator.java
@@ -6,9 +6,9 @@ package me.lokka30.treasury.plugin.bukkit.services.treasury2bukkit;
 
 import me.lokka30.treasury.api.common.event.EventBus;
 import me.lokka30.treasury.api.common.event.EventPriority;
-import me.lokka30.treasury.api.common.services.Service;
-import me.lokka30.treasury.api.common.services.event.ServiceRegisteredEvent;
-import me.lokka30.treasury.api.common.services.event.ServiceUnregisteredEvent;
+import me.lokka30.treasury.api.common.service.Service;
+import me.lokka30.treasury.api.common.service.event.ServiceRegisteredEvent;
+import me.lokka30.treasury.api.common.service.event.ServiceUnregisteredEvent;
 import me.lokka30.treasury.api.economy.EconomyProvider;
 import me.lokka30.treasury.plugin.bukkit.TreasuryBukkit;
 import me.lokka30.treasury.plugin.bukkit.services.ServiceMigrationManager;
@@ -70,7 +70,7 @@ public class T2BServiceMigrator {
                 .completeSubscription());
     }
 
-    private ServicePriority migratePriority(me.lokka30.treasury.api.common.services.ServicePriority priority) {
+    private ServicePriority migratePriority(me.lokka30.treasury.api.common.service.ServicePriority priority) {
         switch (priority) {
             case LOW:
                 return ServicePriority.Low;
